Change the name of th pytest plugin and enhance the support of doctests from other loaders.
