+++ src/importnb/utils/pytest_importnb.py
@@ -4,35 +4,67 @@
 The `AlternativeModule` is reusable.  See `pidgin` for an example.
 """
 
-with __import__("importnb").Notebook():
-    try:
-        from . import testing
-    except:
-        from importnb.utils import testing
-
-import importlib, pytest, abc, pathlib
-
 from importnb import Notebook
+
+import importlib, pytest, abc, pathlib, _pytest, functools
 
 
 def pytest_addoption(parser):
     group = parser.getgroup("general")
     group.addoption("--main", action="store_true", help="Run in the main context.")
+    group.addoption(
+        "--monotonic",
+        action="store_true",
+        help="""Notebook source code has monotonically increasing execution result
+values and a leading Markdown string is the docstring.""",
+    )
 
 
-import _pytest.doctest
+class MonotonicExecution(pytest.File):
+    def collect(self):
+        nb = __import__("json").load(self.fspath.open())
+        yield _pytest.python.Function(
+            assert_execution_order.__name__,
+            self,
+            callobj=functools.partial(assert_execution_order, nb, file=self.fspath),
+        )
 
 
 class AlternativeModule(pytest.Module):
     def _getobj(self):
-        return self.loader(self.parent.config.option.main and "__main__" or None).load(
-            str(self.fspath)
-        )
+        return self.loader(
+            getattr(self.parent.config.option, "main", None) and "__main__" or None
+        ).load(str(self.fspath))
 
     def collect(self):
+        if self.parent.config.option.monotonic:
+            yield from MonotonicExecution.collect(self)
         yield from super().collect()
         if self.parent.config.option.doctestmodules:
+            self.fspath.pyimport = functools.partial(
+                self.fspath.pyimport, modname=self._obj.__name__
+            )
             yield from _pytest.doctest.DoctestModule.collect(self)
+
+
+def assert_execution_order(nb, file=None):
+    shift = 1
+    for id, object in enumerate(object for object in nb["cells"] if object["cell_type"] == "code"):
+        id += shift
+        source = "".join(object["source"])
+        if object["execution_count"] is None:
+            assert (
+                not source.strip()
+            ), """{file} has an unexecuted with the source:\n{source}.""".format(
+                **locals()
+            )
+            shift -= 1
+        else:
+            assert (
+                object["execution_count"] == id
+            ), """{file} has been executed out of order.""".format(**locals())
+
+    return True
 
 
 class NotebookModule(AlternativeModule):
